Test GPG commit signing
